+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/PyramidNetherStructure.java
@@ -4,6 +4,7 @@ import com.google.common.collect.ImmutableList;
 import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Codec;
 import com.telepathicgrunt.repurposedstructures.RepurposedStructures;
+import com.telepathicgrunt.repurposedstructures.world.structures.pieces.GeneralJigsawGenerator;
 import net.minecraft.block.Blocks;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
@@ -20,7 +21,9 @@ import net.minecraft.world.gen.feature.template.*;
 
 public class PyramidNetherStructure extends Structure<NoFeatureConfig> {
     // Special thanks to /r/l-ll-ll-l_IsDisLoss for allowing me to mimic his nether pyramid design!
-    static {
+
+    private static boolean INITIALIZED_POOLS = false;
+    private static void initPools() {
         ImmutableList<StructureProcessor> randomizationList = ImmutableList.of(new RuleStructureProcessor(ImmutableList.of(
                 new RuleEntry(new RandomBlockMatchRuleTest(Blocks.field_235406_np_, 0.04F),
                         AlwaysTrueRuleTest.INSTANCE, Blocks.field_235412_nv_.getDefaultState()),
@@ -43,7 +46,7 @@ public class PyramidNetherStructure extends Structure<NoFeatureConfig> {
     }
 
     public static class Start extends AbstractNetherStructure.AbstractStart {
-        ResourceLocation NETHER_PYRAMID_POOL = new ResourceLocation(RepurposedStructures.MODID,"temples/pyramid_nether");
+        private static ResourceLocation NETHER_PYRAMID_POOL = new ResourceLocation(RepurposedStructures.MODID,"temples/pyramid_nether");
 
         public Start(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox mutableBoundingBox, int referenceIn, long seedIn) {
             super(structureIn, chunkX, chunkZ, mutableBoundingBox, referenceIn, seedIn);
@@ -51,6 +54,10 @@ public class PyramidNetherStructure extends Structure<NoFeatureConfig> {
 
         @Override
         public void init(ChunkGenerator chunkGenerator, TemplateManager structureManager, int chunkX, int chunkZ, Biome biome, NoFeatureConfig NoFeatureConfig) {
+            if(!INITIALIZED_POOLS){
+                initPools();
+                INITIALIZED_POOLS = true;
+            }
             BlockPos blockpos = new BlockPos(chunkX * 16, 35, chunkZ * 16);
             GeneralJigsawGenerator.addPieces(chunkGenerator, structureManager, blockpos, this.components, this.rand, NETHER_PYRAMID_POOL, 1);
             //PyramidFloorPiece.func_207617_a(structureManager, blockpos, this.components.get(0).getRotation(), this.components, random, Blocks.field_235406_np_, NoFeatureConfig);