+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSIglooStructure.java
@@ -1,6 +1,7 @@
 package com.telepathicgrunt.repurposedstructures.world.structures;
 
 import com.mojang.serialization.Codec;
+import com.telepathicgrunt.repurposedstructures.world.structures.pieces.RSIglooPieces;
 import net.minecraft.block.Blocks;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.Rotation;
@@ -15,17 +16,23 @@ import net.minecraft.world.gen.feature.structure.StructureStart;
 import net.minecraft.world.gen.feature.template.TemplateManager;
 
 
-public abstract class AbstractIglooStructure extends AbstractBaseStructure {
-    public AbstractIglooStructure(Codec<NoFeatureConfig> config) {
+public class RSIglooStructure extends AbstractBaseStructure {
+    private final ResourceLocation TOP_PIECE_RL;
+
+    public RSIglooStructure(Codec<NoFeatureConfig> config, ResourceLocation topPieceRL) {
         super(config);
+        TOP_PIECE_RL = topPieceRL;
     }
 
-    public static abstract class AbstractStart extends StructureStart<NoFeatureConfig> {
-        public AbstractStart(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox mutableBoundingBox, int referenceIn, long seedIn) {
+    @Override
+    public IStartFactory<NoFeatureConfig> getStartFactory() {
+        return Start::new;
+    }
+
+    public class Start extends StructureStart<NoFeatureConfig> {
+        public Start(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox mutableBoundingBox, int referenceIn, long seedIn) {
             super(structureIn, chunkX, chunkZ, mutableBoundingBox, referenceIn, seedIn);
         }
-
-        public abstract ResourceLocation getTopPieceResourceLocation();
 
         @Override
         public void init(ChunkGenerator chunkGenerator, TemplateManager structureManager, int chunkX, int chunkZ, Biome biome, NoFeatureConfig NoFeatureConfig) {
@@ -33,7 +40,7 @@ public abstract class AbstractIglooStructure extends AbstractBaseStructure {
             int z = chunkZ * 16;
             BlockPos blockpos = new BlockPos(x, 90, z);
             Rotation rotation = Rotation.values()[this.rand.nextInt(Rotation.values().length)];
-            RSIglooPieces.func_207617_a(structureManager, getTopPieceResourceLocation(), Blocks.PODZOL, blockpos, rotation, this.components, this.rand, IFeatureConfig.NO_FEATURE_CONFIG);
+            RSIglooPieces.func_207617_a(structureManager, TOP_PIECE_RL, Blocks.PODZOL, blockpos, rotation, this.components, this.rand, IFeatureConfig.NO_FEATURE_CONFIG);
             this.recalculateStructureSize();
         }
     }